backend/requirements.txt
fastapi
uvicorn
python-dotenv
azure-ai-inference
azure-core
PyPDF2
python-docx
gunicorn
python-multipart

